// src/api/models/WorkerUnitResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WorkerUnitStatus } from './WorkerUnitStatus';
export type WorkerUnitResponse = {
    id: string;
    status: WorkerUnitStatus;
    statusDescription?: string;
    fileNames?: Array<string>;
};