// src/api/models/WorkerUnitStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum WorkerUnitStatus {
    IN_PROGRESS = 'IN_PROGRESS',
    SUCCESS = 'SUCCESS',
    FAILED = 'FAILED',
    CANCELED = 'CANCELED',
    SYSTEM_ERROR = 'SYSTEM_ERROR',
    NO_REPORT_ERROR = 'NO_REPORT_ERROR',
}